+++ src/backend/adw/nvdialog_dialog_box.c
@@ -25,36 +25,12 @@
 #include "../../nvdialog_assert.h"
 #include "../../nvdialog_macros.h"
 #include "gtk/gtk.h"
+#include "../gtk/nvdialog_gtk.h"
 #include "nvdialog_adw.h"
 
 NvdDialogBox *nvd_dialog_box_adw(const char *title, const char *message,
                                  NvdDialogType type) {
-        NvdDialogBox *dialog = malloc(sizeof(struct _NvdDialogBox));
-        NVD_RETURN_IF_NULL(dialog);
-
-        dialog->content = message;
-        dialog->msg = title;
-        dialog->type = type;
-        dialog->window_handle =
-                adw_message_dialog_new(nvd_get_parent(), title, message);
-        adw_message_dialog_add_response(
-                ADW_MESSAGE_DIALOG(dialog->window_handle), "accept", "Okay");
-        adw_message_dialog_set_default_response(
-                ADW_MESSAGE_DIALOG(dialog->window_handle), "accept");
-
-        switch (type) {
-                case NVD_DIALOG_WARNING:
-                case NVD_DIALOG_ERROR:
-                        adw_message_dialog_set_response_appearance(
-                                ADW_MESSAGE_DIALOG(dialog->window_handle),
-                                "accept", ADW_RESPONSE_DESTRUCTIVE);
-                        break;
-
-                default:
-                case NVD_DIALOG_SIMPLE:
-                        break;
-        }
-        return dialog;
+        return nvd_dialog_box_gtk(title, message, type);
 }
 
 inline void *nvd_dialog_box_get_raw_adw(NvdDialogBox *dlg) {
@@ -63,7 +39,5 @@ inline void *nvd_dialog_box_get_raw_adw(NvdDialogBox *dlg) {
 }
 
 void nvd_show_dialog_adw(NvdDialogBox *dialog) {
-        gtk_window_present(GTK_WINDOW(dialog->window_handle));
-        g_main_context_iteration(NULL, true);
-        /* TODO: Forcefully terminate the Gtk / libadwaita loop. */
+        nvd_show_dialog_gtk(dialog);
 }