+++ src/backend/adw/nvdialog_question_dialog.c
@@ -25,6 +25,7 @@
 #include "../../nvdialog_assert.h"
 #include "../../nvdialog_macros.h"
 #include "nvdialog_adw.h"
+#include "../gtk/nvdialog_gtk.h"
 
 struct NvdQuestionData {
         NvdReply reply;
@@ -43,67 +44,11 @@ inline static void nvd_reply_write_no(NvdReply *reply) {
 
 NvdQuestionBox *nvd_question_adw(const char *title, const char *question,
                                  NvdQuestionButton buttons) {
-        NvdQuestionBox *dialog = malloc(sizeof(struct _NvdQuestionBox));
-        NVD_RETURN_IF_NULL(dialog);
-        dialog->title = (char *)title;
-        dialog->contents = (char *)question;
-        dialog->reply =
-                NVD_REPLY_CANCEL; /* Default reply if no other was given */
-
-        dialog->window_handle = (AdwMessageDialog *)adw_message_dialog_new(
-                nvd_get_parent(), dialog->title, dialog->contents);
-
-        NVD_CHECK_INTERNAL(dialog->window_handle, dialog, NULL);
-
-        switch (buttons) {
-                case NVD_YES_NO:
-                        adw_message_dialog_add_response(
-                                ADW_MESSAGE_DIALOG(dialog->window_handle),
-                                "cancel", "Cancel");
-                        break;
-                case NVD_YES_NO_CANCEL:
-                        adw_message_dialog_add_response(
-                                ADW_MESSAGE_DIALOG(
-                                        (GtkWidget *)dialog->window_handle),
-                                "cancel", "Cancel");
-                        adw_message_dialog_add_response(
-                                ADW_MESSAGE_DIALOG(
-                                        (GtkWidget *)dialog->window_handle),
-                                "reject", "No");
-                        break;
-                case NVD_YES_CANCEL:
-                        adw_message_dialog_add_response(
-                                ADW_MESSAGE_DIALOG(
-                                        (GtkWidget *)dialog->window_handle),
-                                "reject", "No");
-                        break;
-                default:
-                        return NULL;
-        }
-        adw_message_dialog_add_response(
-                ADW_MESSAGE_DIALOG((GtkWidget *)dialog->window_handle),
-                "accept", "Okay");
-        /* Just for better looks. */
-        adw_message_dialog_set_response_appearance(
-                ADW_MESSAGE_DIALOG((GtkWidget *)dialog->window_handle),
-                "accept", ADW_RESPONSE_SUGGESTED);
-        return dialog;
+        return nvd_question_gtk(title, question, buttons);
 }
 
 NvdReply nvd_get_reply_adw(NvdQuestionBox *box) {
-        g_signal_connect_swapped((GtkWidget *)box->window_handle,
-                                 "response::accept",
-                                 G_CALLBACK(nvd_reply_write_ok), &box->reply);
-        g_signal_connect_swapped((GtkWidget *)box->window_handle,
-                                 "response::reject",
-                                 G_CALLBACK(nvd_reply_write_no), &box->reply);
-        g_signal_connect_swapped(
-                (GtkWidget *)box->window_handle, "response::cancel",
-                G_CALLBACK(nvd_reply_write_cancel), &box->reply);
-
-        gtk_window_present(GTK_WINDOW(box->window_handle));
-        g_main_context_iteration(NULL, true);
-        return box->reply;
+        return nvd_get_reply_gtk(box);
 }
 
 void *nvd_dialog_question_get_raw_adw(NvdQuestionBox *dlg) {